Lab6/requirements.txt
asgiref==3.4.1
Django==2.1.15
djangorestframework==3.11.0
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-pyodbc-azure-2019==2.1.0.0
pyodbc==4.0.32
pytz==2021.3
sqlparse==0.4.2
